basic version extraction and writeback
this is surprisingly complex because it can only be written back to a branch
(because tags cannot be committed to in git), but only when the version was
originally extracted from the VERSION file (because pom.xml extraction will
not allow updates).

also enable javascript strict mode, which reports errors for all those
"features" of the language that should never have existed in the first place.