add project selection page; refactor ref handling
the project selection page shouldn't be necessary; in principle, the user
should have a direct link from GitLab which includes the project name.
however it us useful as a stopgap until that's implemented, and may be
useful as a fallback even afterwards.

improved ref handling now allows /api/repo/selected-refs to be fast, which
in practice means that the metadata page is more usable.