don't throw 400s when returning after PushTask is finished
when returning after PushTask is finished, the push is invalidated so
the user can try it again (also needed for exception handling). this
used to delete the project to clean up, but that doesn't currently work
and just reports an HTTP 400 from GitLab. removing the deleteProject()
means dead projects in the archive, but that's currently unavoidable
anyway.

the correct fix is moving the project to the `archive` namespace after
a successful push. unfortunately that is currently an admin-only
operation.