refactor license selection
it now checks for exactly the conditions that are needed in PushTask to
update the LICENSE file. the only check beyond that is that if multiple
licenses have been detected, we only allow this to be overridden on the
single-license page if it had been previously overridden on the
multiple-licenses page.