don't show error message on ProjectCompletedException
users tend not to read error messages, so it just looks like returning
broke the webapp. it didn't, but that's hard to tell.

the same problem might happen for the other exceptions as well, but
these are somewhat hard to trigger. for the ProjectCompletedException,
all it takes is clicking the back button after archiving...