delay item creation until after access rights page
this means the item can be directly created with the correct access
rights etc.

if the create-and-move workflow is implemented, we can still delete the
archived data after pushing it. this turns the button on overview.html
into a notmal "next" button that can be undone. clicking the access
rights button is final, however. label it "complete archiving process"
to make sure user knows that.