renew deployment SSH key
having the key in the credentials submodule means that if the user
triggering a build doesn't have access to that repo, the deployment
will fail. this can only happen for users added to this project but
not to the group: the build inherits the premissions of the user that
triggered it, and so wouldn't have read access to the credentials
repo in this specific case.