correctly handle refs along other branches
when a branch A was a direct ancestor of B, processing A after B would
blindly add A to the stack, triggering the "processed twice" exception.
the correct thing to do it to just not process A at all in this case,
because its rewrite result is already known.

removing the "pocessed twice" exception is *not* an option; it is
necessaray to guarantee that rewriting never gives different results
for the same commit in different branches.